feat: reutilize code
